moved locations of validator classes and updated namespace and use statements
