+++ src/Validators/ISO8601ValidatorTest.php
@@ -2,7 +2,7 @@
 
 use PHPUnit\Framework\TestCase;
 
-use PDGA\DataObjects\ISO8601Validator;
+use PDGA\DataObjects\Validators\ISO8601Validator;
 
 final class ISO8601ValidatorTest extends TestCase
 {
