+++ src/Validators/IntValidatorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PDGA\DataObjects;
+namespace PDGA\DataObjects\Validators;
 
 use PHPUnit\Framework\TestCase;
