+++ src/Validators/MinLengthValidator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PDGA\DataObjects;
+namespace PDGA\DataObjects\Validators;
 
 /**
  * Validates minimum length of strings, integers and arrays.
